% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_methods.R
\name{accessors}
\alias{accessors}
\alias{path,ribo-method}
\alias{experiments,ribo-method}
\alias{format_version,ribo-method}
\alias{reference,ribo-method}
\alias{length_min,ribo-method}
\alias{length_max,ribo-method}
\alias{left_span,ribo-method}
\alias{right_span,ribo-method}
\alias{metagene_radius,ribo-method}
\alias{length_offset,ribo-method}
\alias{has_metadata,ribo-method}
\alias{experiment_info,ribo-method}
\alias{transcript_info,ribo-method}
\alias{alias_hash,ribo-method}
\alias{original_hash,ribo-method}
\title{Ribo Object Accessors}
\usage{
\S4method{path}{ribo}(object)

\S4method{experiments}{ribo}(object)

\S4method{format_version}{ribo}(object)

\S4method{reference}{ribo}(object)

\S4method{length_min}{ribo}(object)

\S4method{length_max}{ribo}(object)

\S4method{left_span}{ribo}(object)

\S4method{right_span}{ribo}(object)

\S4method{metagene_radius}{ribo}(object)

\S4method{length_offset}{ribo}(object)

\S4method{has_metadata}{ribo}(object)

\S4method{experiment_info}{ribo}(object)

\S4method{transcript_info}{ribo}(object)

\S4method{alias_hash}{ribo}(object)

\S4method{original_hash}{ribo}(object)
}
\arguments{
\item{object}{A Ribo Object}
}
\description{
A list of accessors for the different slots of a `Ribo` object
}
