% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ribo.R
\name{create_ribo}
\alias{create_ribo}
\title{Creates an S4 object of class "ribo"}
\usage{
create_ribo(name, rename = NULL)
}
\arguments{
\item{name}{The path to the .ribo file}

\item{rename}{A function that renames the original transcript or an already generated
character vector of aliases}
}
\value{
Returns an S4 object of class "ribo" containing a path to the HDF5 file,
        various attributes in the root folder, and information about the transcripts 
        such as names and lengths
}
\description{
\code{\link{create_ribo}} creates a "ribo" object. It creates a path, extracts the root folder attributes,
and provides information about the reference transcript names and lengths
}
\details{
An important option is the param 'rename' which allows the user to nickname long
transcript names. For the appris human transcriptome, a default function \code{\link{rename_default}}
has been provided. In subsequent calls of certain functions, the user can make use of these renamed
references with the 'alias' parameter.

This object is required as an argument for almost all of the functions in this package, and all of the
functions in this package can accept the returned object of this function. This object is not meant to
be modified or changed by the user. It is meant to serve as an intermediary between the .ribo file and
an R environment by creating an object that holds pertinent information.

The information stored in this object include the .ribo file path, the list of experiments,
the format version, the reference model, the maximum read length, the minimum read length, the left span,
the right span, and other information about the transcript information.

Some of the subsequent function calls avoid direct usage of the information stored in the
.ribo object to prevent any accidental error. However, certain variables in the returned object, such
as the path, are required to make use of the additional functionality in this package.
}
\examples{
#generate a ribo object with transcript nicknames/aliases
file.path <- system.file("extdata", "HEK293_ingolia.ribo", package = "ribor")
sample <- create_ribo(file.path, rename = rename_default )
}
\seealso{
If a ribo object is already generated but aliases want to be added or updated, use the
\code{\link{set_aliases}} function.
}
