% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_attributes}
\alias{get_attributes}
\title{Helper method that reads the attributes}
\usage{
get_attributes(ribo.object)
}
\description{
get_attributes retrieves all attributes except for time
}
