% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_internal_region_lengths}
\alias{get_internal_region_lengths}
\title{Returns the overall length of each region with UTR Junctions}
\usage{
get_internal_region_lengths(ribo.object, alias = FALSE)
}
\arguments{
\item{ribo.object}{A 'Ribo' object}

\item{alias}{Option to return the transcript names as aliases}
}
\value{
A data.frame of the region lengths
}
\description{
The function {\code{\link{get_internal_region_coordinates}}} retrieves the 
lengths for the UTR5, UTR5 Junction, CDS, UTR3 Junction,
and UTR3 regions of every transcript.
}
\examples{
# generate a ribo object 
file.path <- system.file("extdata", "HEK293_ingolia.ribo", package = "ribor")
sample <- Ribo(file.path, rename = rename_default)

# get the region lengths
region_lengths <- get_internal_region_lengths(sample, alias = TRUE)

}
