% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_functions.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieves the metadata of an experiment}
\usage{
get_metadata(ribo.object, name)
}
\arguments{
\item{ribo.object}{S3 class of object ribo}

\item{name}{The name of the experiment}
}
\value{
If the experiment is valid, a list of 4 elements providing 
information about the cell_line, digestion_duration, digestion_enzyme, and 
source link is returned.

If the name is not found, then a list of attributes describing the root file 
is provided.
}
\description{
\code{\link{get_metadata}} provides information on the cell_line, digestion_duration, 
digestion_enzyme, and a website link to the source if the experiment is found in the .ribo
file. If the experiment is not found, then the attributes of the root .ribo file is returned instead.
}
\seealso{
\code{\link{ribo}} to generate the necessary ribo.object parameter
}
