% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_read_lengths}
\alias{get_read_lengths}
\title{Helper method that retrieves the minimum and maximum read lengths}
\usage{
get_read_lengths(ribo.object)
}
\description{
get_read_lengths finds the minimum and maximum read lengths of the .ribo file
}
