% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_reference_lengths}
\alias{get_reference_lengths}
\title{Information about the reference names}
\usage{
get_reference_lengths(ribo.object)
}
\arguments{
\item{ribo.object}{S3 object with class "ribo"}
}
\value{
Returns a data table of reference transcript names and lengths
}
\description{
get_reference_lengths() takes the "ribo" S3 object and reads in the
reference names and reference transcripts
}
