% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_reference_names}
\alias{get_reference_names}
\title{Information about the reference names}
\usage{
get_reference_names(ribo.object)
}
\arguments{
\item{ribo.object}{S3 object with class "ribo"}
}
\value{
Returns a list of the reference names
}
\description{
Information about the reference names
}
