% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{help_rnaseq}
\alias{help_rnaseq}
\title{helper method for get_rnaseq}
\usage{
help_rnaseq(handle, result, experiment.list, transcript.list, rnaseq.list,
  has_missing)
}
\description{
helper method for get_rnaseq
}
