% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{is.ribo}
\alias{is.ribo}
\title{Checks for a valid 'ribo' object}
\usage{
is.ribo(ribo.object)
}
\arguments{
\item{ribo.object}{Object in question}
}
\value{
A boolean indicating whether or not a given object is a valid 'ribo' object
}
\description{
The function \code{\link{is.ribo}} checks whether or not a given ribo object
is valid.
}
\examples{
file.path <- system.file("extdata", "HEK293_ingolia.ribo", package = "ribor")
sample <- create_ribo(file.path, rename = rename_default)

is.ribo(sample)
}
