% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_count_functions.R
\name{plot_length_distribution}
\alias{plot_length_distribution}
\title{Plots the length distribution}
\usage{
plot_length_distribution(x, region, range.lower, range.upper, experiments,
  percentage = TRUE, title = "Length Distribution")
}
\arguments{
\item{x}{either an S3 "ribo" object or a data.table}

\item{region}{the region of interest}

\item{range.lower}{a lower bounds for a read length range}

\item{range.upper}{an upper bounds for a read length range}

\item{experiments}{a list of experiment names}

\item{percentage}{logical value that, if TRUE, presents the data as a percentage}

\item{title}{a title for the generated plot}
}
\description{
The function \code{\link{plot_region_counts}} can take either a data.table
or a "ribo" object to generate a line graph of the length distributions from
range.lower to range.upper.
}
\details{
The param 'percentage' will plot the percentages of each length relative
to the total sum of the read length range provided by param 'range.lower'
and 'range.upper'. When percentage is set to FALSE, the total count of each
read length is plotted.

When given a "ribo" object, \code{\link{plot_length_distribution}} calls
\code{\link{get_region_counts}} to retrieve the necessary information
for plotting. This option is in the case that a data.table of the
region count information is not required.

The user can instead provide a data.table with the same structure as the
output of the \code{\link{get_region_counts}} function where the 'transcript'
parameter is set to FALSE and 'length' parameters is the default value of
TRUE. This also means that the remaining parameters of the
\code{\link{plot_length_distribution}} function are not necessary. The run
time becomes substantially faster when \code{\link{plot_region_counts}} is
given the direct data.table to plot. Note that there is no manipulation by
this function on the data.table, making this input option more error prone.
}
\seealso{
\code{\link{get_region_counts}} to generate a data.table that can
be provided as input,
\code{\link{ribo}} to create a ribo.object that can be provided as input
}
