% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_functions.R
\name{plot_metagene}
\alias{plot_metagene}
\title{Plots the metagene coverage data}
\usage{
plot_metagene(x, site, range.lower, range.upper, experiments,
  title = "Metagene Site Coverage")
}
\arguments{
\item{x}{ribo.object or data.table}

\item{site}{"start" or "stop" site}

\item{range.lower}{lower bound of the read length, inclusive}

\item{range.upper}{upper bound of the read length, inclusive}

\item{experiments}{list of experiments}

\item{title}{title of the generated plot}
}
\description{
The function \code{\link{plot_metagene}} plots the metagene site coverage,
separating by experiment.
}
\details{
If a data.table is provided as param 'x', then the only additional parameter
is the optional title' parameter for the generated plot. If a ribo.object is
provided as param 'x', the rest of the parameters listed are necessary.

When given a ribo class object, the \code{\link{plot_metagene}} function
generates a data.table by calling the \code{\link{get_tidy_metagene}}
function, so the run times in this case will be mostly comprised of a call
to the \code{\link{get_metagene}} function.

This function uses ggplot in its underlying implementation.
}
\examples{
#a common use case is to directly pass in the ribo object file as param 'x'
#generate the ribo object by loading in a ribo function and calling the
#\\code{\\link{ribo}} function
\donttest{
sample <- ribo("sample.ribo")
}


#assuming sample contains the following experiments
\donttest{
experiments <- c("HeLa_1", "HeLa_2", "HeLa_3")
}


#plot the metagene start site coverage for all experiments in 'sample.ribo'
#from read length 15 to 40
\donttest{
plot_metagene(x = sample,
              site = "start",
              range.lower = 15,
              range.upper = 40,
              experiments = experiments)
}


#Note that the site, range.lower, range.upper, and experiments are only
#necessary if a ribo object is being passed in as param 'x'. If a ribo
#object is passed in, then the param 'experiments' will be set to all of
#the experiments by default.

#If a data.table is passed in, then the \\code{\\link{plot_metagene}} function
#does not need any other information. All of the elements of the data.table
#will be used, assuming that it contains the same column names and number of
#columns as the output from \\code{\\link{get_tidy_metagene}}

#gets the metagene start site coverage from read length 15 to 40
#note that the data must be summed across transcripts and read lengths
#for the plot_metagene function
\donttest{
data <- get_tidy_metagene(sample,
                          site = "start",
                          range.lower = 15,
                          range.upper = 40)
}

\donttest{
plot_metagene(data)
}


}
