% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_object.R
\name{ribo}
\alias{ribo}
\title{Creates an S3 object of class "ribo"}
\usage{
ribo(name)
}
\arguments{
\item{name}{The path to the .ribo file}
}
\value{
Returns a list containing a handle to the HDF5 file,
       various attributes in the root folder, and
       information about the transcripts such as
       names and lengths
}
\description{
\code{\link{ribo}} creates an S3 object. It creates a handle, extracts the root folder attributes,
and provides information about the reference transcript names and lengths
}
\details{
This object is required as an argument for almost all of the functions in this package, and all of the
functions in this package can accept the returned object of this function. This object is not meant to 
be modified or changed by the user. It is meant to serve as an intermediary between the .ribo file and 
an R environment by creating an object that holds pertinent information.

The information stored in this object include the .ribo file handle, the list of experiments,
the format version, the reference model, the maximum read length, the minimum read length, the left span,
the right span, and other information about the transcript information.

Some of the subsequent function calls avoid direct usage of the information stored in the 
.ribo object to prevent any accidental error. However, certain variables in the returned object, such 
as the handle, are required to make use of the additional functionality in this package.
}
