% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_object.R
\name{riboFopen}
\alias{riboFopen}
\title{Creates an S3 object of class "ribo"}
\usage{
riboFopen(name)
}
\arguments{
\item{name}{The filename of the HDF5 file}
}
\value{
Returns a list containing a handle to the HDF5 file,
       various attributes in the root folder, and
       information about the transcripts such as
       names and lengths
}
\description{
riboFopen() creates an S3 object. It creates a handle, extracts the root folder attributes,
and provides information about the reference transcript names and lengths
}
