% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribor.R
\docType{package}
\name{ribor}
\alias{ribor}
\alias{ribor-package}
\title{ribor: A package for reading .ribo files}
\description{
The 'ribor' package offers a suite of reading functions for any 
of the datasets present in a .ribo file and also provides some rudimentary
plotting functions.
}
