% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_reference_names}
\alias{get_reference_names}
\title{Retrieves a list of reference names}
\usage{
get_reference_names(ribo.object)
}
\arguments{
\item{ribo.object}{A 'ribo' object-=09}
}
\value{
a list of the reference names
}
\description{
Gets a list of reference names by reading directly from the .ribo file
}
\examples{
#generate a ribo object with transcript nicknames/aliases
file.path <- system.file("extdata", "HEK293_ingolia.ribo", package = "ribor")
sample <- create_ribo(file.path)

#get the reference names
names <- get_reference_names(sample)
}
