% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_object.R
\name{print.ribo}
\alias{print.ribo}
\title{Printing "ribo" objects}
\usage{
\method{print}{ribo}(x, ...)
}
\arguments{
\item{x}{object of class "ribo"}

\item{...}{additional arguments that have no function yet}
}
\description{
Print a ribo object
}
\details{
This method retrieves any user-relevant information about the ribo object 
and neatly prints it out. The information is extracted from the "ribo" class
object, turned into a data.table. The entries are then converted to a character,
customly formatted, and printed.
}
