% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_class.R, R/ribo_methods.R
\docType{class}
\name{Ribo-class}
\alias{Ribo-class}
\alias{show,Ribo-method}
\title{Class "Ribo"}
\usage{
\S4method{show}{Ribo}(object)
}
\arguments{
\item{object}{A 'Ribo' object}
}
\value{
A 'Ribo' object
}
\description{
An S4 class to be used with the ribor package
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Displaying the file contents
}}

\section{Slots}{

\describe{
\item{\code{path}}{A path to the Ribo file of interest}

\item{\code{experiments}}{A character vector of experiment names in the file}

\item{\code{format.version}}{The format version of the ribo file}

\item{\code{reference}}{The reference transcriptome used in the ribo file}

\item{\code{length.min}}{The minimum read length of the data in the file}

\item{\code{length.max}}{The maximum read length of the data in the file}

\item{\code{left.span}}{Left span of the junction regions}

\item{\code{right.span}}{Right span of the junction regions}

\item{\code{length.offset}}{Length offset of all transcripts}

\item{\code{has.metadata}}{Value denoting whether the root ribo file has metadata}

\item{\code{experiment.info}}{data.frame of information on the experiments}

\item{\code{transcript.info}}{Hash of the lengths and offsets of each transcript}

\item{\code{alias.hash}}{Hash that goes from alias to original transcript name}

\item{\code{original.hash}}{Hash that goes from original to alias transcript}
}}

\examples{
file.path <- system.file("extdata", "sample.ribo", package = "ribor")
sample <- Ribo(file.path)

show(sample)
}
\seealso{
\code{\link{Ribo}} to create a ribo file
}
